challenges/misc/bottle-flip-challenge/infra/requirements.txt
piexif==1.1.3
Pillow==11.1.0
